% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_heading_to_table.R
\name{add_heading_to_table}
\alias{add_heading_to_table}
\title{Add Heading to Table.}
\usage{
add_heading_to_table(table, heading = "Heading", fontsize = 18)
}
\arguments{
\item{table}{gridExtra::tableGrob // The table to add the heading to.}

\item{heading}{string // The heading to add.}

\item{fontsize}{numeric // The size of the font of the heading.}
}
\value{
The gridExtra::tableGrob table with the heading at the top.
}
\description{
Append a heading to the top of a gridExtra::tableGrob table.
}
