% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ansi_colours}
\alias{ansi_colours}
\title{ANSI Colour Data}
\format{
## `ansi_colours`
A named character array with length 8.
}
\usage{
ansi_colours
}
\description{
Each of the standard ANSI colour codes in 'ready-to-print' format.
}
\keyword{datasets}
