% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_tomR_theme_to_rstudio.R
\name{apply_tomR_theme_to_rstudio}
\alias{apply_tomR_theme_to_rstudio}
\title{Apply custom tomR theme to RStudio.}
\usage{
apply_tomR_theme_to_rstudio()
}
\description{
Install and apply the "official" tomR theme!
}
