% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{ask_from_path}
\alias{ask_from_path}
\title{Ask which file to load from a Path.}
\usage{
ask_from_path(path)
}
\arguments{
\item{path}{The path to the directory where your files are located.}
}
\description{
Look at all of the files in a directory and be prompted to choose one of them to load.
}
