% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{b_by_cond}
\alias{b_by_cond}
\title{Boundary By Condition}
\usage{
b_by_cond(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\value{
A vector of length t_vec containing b from prms_model based on one_cond.
}
\description{
Function to calculate b when it differs by condition.
}
