% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{b_fixed}
\alias{b_fixed}
\title{Fixed Boundary}
\usage{
b_fixed(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\value{
A vector of length t_vec containing b from prms_model.
}
\description{
Function to calculate b when it is fixed across conditions.
}
