% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combied_sat_measures.R
\name{bis}
\alias{bis}
\title{Balanced integration score}
\usage{
bis(rt, pc)
}
\arguments{
\item{rt}{vector-like. The response time vector.}

\item{pc}{vector-like. The 'percent correct' vector.}

\item{places}{numeric, default = 3. Value to round to. Pass in 0 or FALSE to not round.}
}
\description{
Calculate BIS for a provided set of rts and pcs.
}
