% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{build_tomR}
\alias{build_tomR}
\title{Build tomR.}
\usage{
build_tomR(version = "debug")
}
\arguments{
\item{version}{major, minor, patch, debug // Type of increment to perform.}
}
\description{
Build this package and iterate its version.
}
