% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{check_git}
\alias{check_git}
\title{Check (g)it!}
\usage{
check_git(repo, verbose = F)
}
\arguments{
\item{repo}{The name of the Github repository to check.}

\item{verbose}{If true print lots of information to the console.}
}
\value{
True if the latest github version and installed version match.
}
\description{
Check the current version number of the entered repo.
}
