% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_function.R
\name{coffee_function}
\alias{coffee_function}
\title{Coffee Function}
\usage{
coffee_function(brand, caffeine, size)
}
\arguments{
\item{brand}{What brand of coffee "Character"}

\item{caffeine}{How many mg of caffeine is in it "Numeric"}

\item{size}{What size the coffee cup is}
}
\description{
Coffee Function
}
\examples{
coffee_function("kirkland", 100, "large")
}
