% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Cohen's D (Paired T-Test)}
\usage{
cohens_d(x, y)
}
\arguments{
\item{x}{Numeric vector of data from population x.}

\item{y}{Numeric vector of data from population y.}
}
\value{
The calculated d parameter, automatically correct for Pearson's r as suggested by Cohen.
}
\description{
Calculate Cohen's d using the t-statistic from the provided variables.
}
