% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridExtra.R
\name{colour_blue_to_orange}
\alias{colour_blue_to_orange}
\title{Colour Blue to Orange.}
\usage{
colour_blue_to_orange(table)
}
\arguments{
\item{table}{The table to apply colouring to.}
}
\value{
A gridExtra::tableGrob containing the values from the matrix.
}
\description{
Apply a blue to orange gradient to values in this table showing
how positive or negative they are.
}
