% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_printer.R
\name{colour_printer}
\alias{colour_printer}
\title{Colour Printer.}
\usage{
colour_printer(..., colour = "red", sep = " ", lb = TRUE)
}
\arguments{
\item{...}{Cohered to String // Content to print to the console.}

\item{colour}{string // black, red, green, blue, pink, lavender, or yellow.}

\item{sep}{string // Separator to use between arguments.}

\item{lb}{bool // Should an empty line be printed after the message?}
}
\description{
Print to the console in colour.
}
