% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridExtra.R
\name{convert_matrix_to_table}
\alias{convert_matrix_to_table}
\title{Convert Matrix to Table}
\usage{
convert_matrix_to_table(matrix)
}
\arguments{
\item{matrix}{The matrix to convert.}
}
\value{
A gridExtra::tableGrob containing the values from the matrix.
}
\description{
Convert a matrix to a gridExtra::tableGrob.
}
\note{
# returns the table with an aesthetic similar to the ggplot theme_classic
aesthetics. You cannot control or change those details here.
}
