% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridExtra.R
\name{convert_table_to_matrix}
\alias{convert_table_to_matrix}
\title{Convert Table to Matrix}
\usage{
convert_table_to_matrix(table)
}
\arguments{
\item{table}{The gridExtra::tableGrob to convert.}
}
\value{
A matrix of the cell values from the provided table.
}
\description{
Convert a gridExtra::tableGrob to a matrix.
}
