% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{cran_it}
\alias{cran_it}
\title{CRAN It!}
\usage{
cran_it(package)
}
\arguments{
\item{package}{The name of the package to try and load.}
}
\description{
Try to load a package, and if you don't have it try to install it from CRAN instead.
}
