% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_placeholder_matrix.R
\name{create_matrix_of_random_values}
\alias{create_matrix_of_random_values}
\title{Create a Matrix of Random Values.}
\usage{
create_matrix_of_random_values(len = 3)
}
\arguments{
\item{len}{Length of the matrix sides.}
}
\value{
A matrix containing random valuesof size len.
}
\description{
Create a placeholder matrix with random values in it, where
those above the diagonal are positive and those below are negative.
}
