% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_placeholder_matrix.R
\name{create_placeholder_matrix}
\alias{create_placeholder_matrix}
\title{Create a Placeholder Matrix.}
\usage{
create_placeholder_matrix(len = 3, diags = c("below", "above"))
}
\arguments{
\item{len}{Length of the matrix sides.}

\item{diags}{list // What to place above and below the diagonal line.}
}
\value{
A matrix containing the values provided in diags of size len.
}
\description{
Create a placeholder matrix to see how one will look.
}
