% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{create_sig_matrix}
\alias{create_sig_matrix}
\title{Create Significance Matrix}
\usage{
create_sig_matrix(data, dim, diag, over, var, alpha = 0.05)
}
\arguments{
\item{data}{data.frame // The data.}

\item{dim}{string // The variable to form the matrix from, i.e. the 'sides'.}

\item{diag}{string // The variable to vary above and below the diagonal line.}

\item{over}{string // The variable to loop over while counting number of significant tests.}

\item{var}{string // The variable to perform t-tests on.}

\item{alpha}{numeric, default = 0.05 // The alpha value to use for the t-tests.}
}
\description{
Create a matrix counting the number of positive t-tests.
}
