% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{dm}
\alias{dm}
\title{Direct Message.}
\usage{
dm(..., sep = "", lb = TRUE)
}
\arguments{
\item{...}{cohercable to string // Content to print to the console.}

\item{sep}{string // Separator to use between arguments.}

\item{lb}{bool // Should an empty line be printed after the message?}
}
\description{
Wrapper for printing to the console using message.
}
