% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_functions.R
\name{do_all_levels_match}
\alias{do_all_levels_match}
\title{Do All Levels Match?}
\usage{
do_all_levels_match(a, b, vars, ..., verbose = F)
}
\arguments{
\item{a}{data.frame // One of the objects to compare.}

\item{b}{data.frame // The other object to compare.}

\item{vars}{string or list of strings // The list of variables to observe.}

\item{...}{string // Additional variables to observe.}

\item{verbose}{bool, default = F // Should we print stuff to the console?}
}
\value{
A bool describing if every level of every variable is the same between a and b.
}
\description{
Check every level of the variables provided in object a and b are the same.
}
\note{
Because of the placement of ... you can pass an entire list to vars,
such as c('rt','pc','ID'), or you can write that list directly into the function
such as compare_levels_of_variables(a, b, 'rt', 'pc', 'ID') and the function will
work the exact same way.
}
