% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_content.R
\name{get_cell_content}
\alias{get_cell_content}
\title{Get Cell Content}
\usage{
get_cell_content(table, x, y, ignore_headings = T)
}
\arguments{
\item{table}{gridExtra::tableGrob // The table to search the contents of.}

\item{x}{numeric // The x position (row) to look at.}

\item{y}{numeric // The y position (column) to look at.}

\item{ignore_headings}{bool // Should the x and y coordinates be adjusted to
ignore the heading row and row-number coolumn?}
}
\description{
Find the content of the specified cell of a table.
}
