% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{ktt}
\alias{ktt}
\title{Kill the Twin}
\usage{
ktt(rt, score)
}
\arguments{
\item{rt}{Numeric Vector // A vector of response times.}

\item{score}{Numeric Vector // A vector of scores representing correct or incorrect.}
}
\value{
Returns a data frame of survivors after performing "kill the twin".
}
\description{
Perform a "kill the twin" procedure (Eriksen, 1988).
}
