% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai.R
\name{lai}
\alias{lai}
\title{Log A Index}
\usage{
lai(rt, pc, s = 0.305, A = 5.294, max_p = 0.995)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores. The average PC score for
this block can be taken as p, the probabilty of a correct response.}

\item{s}{The minimum response time required for accuracy to be greater than chance.
This is equivilent to non-decision time, and as such the default value provided
is simply a common value for t0 plus a marginal amount.}

\item{A}{Asymptotic accuracy, aka, the maximum value logit(p) can ever take.
The default value provided (5.294) is marginally higher than logit(p) would be at an
accuracy score of 99.5\%.}

\item{max_p}{Decimal indicating the maximum possible probability of a correct response.
Dennis & Evans (1996) propose that this value should be arbitrarily set to 99.5\%.}
}
\value{
A vector of scores of the same length as the passed vectors.
}
\description{
Estimate Log A Index (Dennis & Evans, 1996) for the given block.
}
