% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{lai_block}
\alias{lai_block}
\title{Log A Index (Block)}
\usage{
lai_block(rt, pc, s = 0.2, A = log(999))
}
\arguments{
\item{rt}{Mean response time of the given block.}

\item{pc}{Proportion correct for the given block.}

\item{s}{The minimum response time required for accuracy to be greater than chance.}

\item{A}{Asymptotic accuracy, aka, the maximum value logit(p) can ever take.}
}
\description{
Estimate Log A Index (Dennis & Evans, 1996) for the given averages.
}
