% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveload.R
\name{lifesaver}
\alias{lifesaver}
\title{Lifesaver}
\usage{
lifesaver(
  .object,
  file,
  warning = "File already found, I refuse to save. Do so manually if you must."
)
}
\arguments{
\item{.object}{The R object to try and save.}

\item{file}{The file path you want to save to.}

\item{warning}{The warning message to display if a conflict occurs.}
}
\description{
Wrapper to prevent accidentally saving over a file that already exists.
}
