% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{lisas}
\alias{lisas}
\title{Linear integration score}
\usage{
lisas(rt, pc)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}
}
\description{
Calculate LISAS.
}
