% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_do.R
\name{load_or_do}
\alias{load_or_do}
\title{Load a File or do Something.}
\usage{
load_or_do(path, do, quiet = F)
}
\arguments{
\item{path}{The path to the file you want to load, or save to if nothing is loaded.}

\item{do}{Function to perform if no file can be loaded.}

\item{quiet}{Suppress print and message commands. Default to false.}
}
\description{
Attempt to load file at this path, or save to this path if something was done.
}
