% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{meanify}
\alias{meanify}
\title{Meanify}
\usage{
meanify(data, ..., rt.col = "RT", score.col = "Score")
}
\arguments{
\item{data}{The data.frame to meanify.}

\item{...}{The columns to group by.}

\item{rt.col}{The current name of the RT column. Will be renamed 'RT'.}

\item{score.col}{The current name of the score column. Will become the 'PC' (percentage correct) column.}
}
\description{
Group and summarise a data.frame into a meanified version.
}
