% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{muc_by_cond}
\alias{muc_by_cond}
\title{Drift Rate By Condition}
\usage{
muc_by_cond(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\description{
Function to calculate muc when it differs by condition.
}
