% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{muc_fixed}
\alias{muc_fixed}
\title{Fixed Drift Rate}
\usage{
muc_fixed(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\value{
A vector of length t_vec containing muc from prms_model.
}
\description{
Function to calculate muc when it is fixed across conditions.
}
