% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{nested_progress}
\alias{nested_progress}
\title{Nested Progress (of some loops).}
\usage{
nested_progress(..., dps = 0)
}
\arguments{
\item{...}{Any number of pairs of the format c(current, max). Each listed pair will become one percentage display in the console.}

\item{dps}{Number of decimal places to round the percentage to.}
}
\description{
cat the current progress of a set of nested loops. Always uses a \\r call to clear the line and replace the current percentage.
}
