% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{non_dec_fixed}
\alias{non_dec_fixed}
\title{Fixed Non-Decision Time}
\usage{
non_dec_fixed(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\value{
A vector of length t_vec containing non_dec from prms_model.
}
\description{
Function to calculate non-decision time when it is fixed across conditions.
}
