% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{The vector to normalize.}
}
\value{
A vector of length x with each value normalized.
}
\description{
Re-scale a vector of any length to the range 0 to 1, aka, normalize it.
}
