% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{pride}
\alias{pride}
\title{Pride Print.}
\usage{
pride(..., sep = " ", lb = TRUE)
}
\arguments{
\item{...}{Cohered to String // Content to print to the console.}

\item{sep}{string // Separator to use between arguments.}

\item{lb}{bool // Should an empty line be printed after the message?}
}
\description{
Print to the console in a cycle of colours, evoking the rainbow.
}
