% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Progress (of this loop).}
\usage{
progress(x, y, dps = 0, colour = "white")
}
\arguments{
\item{x}{Numeric variable indicating the current state of the loop.}

\item{y}{Numeric variable indicating the end of the loop.}

\item{dps}{Number of decimal places to round the percentage to.}

\item{colour}{The colour to print the percentage in. Must be available in \code{\link{ansi_colours}}.}
}
\description{
cat the current progress of the loop as a percentage of x over y.
Always uses a \\r call to clear the line and replace the current percentage.
}
