% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd}
\alias{psd}
\title{Percentage Standard Deviation}
\usage{
psd(value, dps = 3, verbose = F)
}
\arguments{
\item{value}{numeric. Percentage from which the SD should be calculated.}

\item{dps}{numeric, default = 3 // Value to round to. Pass in 0 or FALSE to not round.}

\item{verbose}{bool, default = false // Should stuff be printed to the console or not.}
}
\description{
Calculate the standard deviation of a percentage/mean value.
}
