% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{quick.group.sim}
\alias{quick.group.sim}
\title{Quick Group Simulation}
\usage{
quick.group.sim(n = 100, N = 20)
}
\arguments{
\item{n}{The number of trials to simulate per subject}

\item{N}{the number of subjects to simulate.}
}
\value{
Returns a data frame with columns 'ID','RT', and 'Score'.
}
\description{
Perform a very quick simulation of one subject using dRiftDM.
}
