% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{quick.subject.sim}
\alias{quick.subject.sim}
\title{Quick Subject Simulation}
\usage{
quick.subject.sim(n = 100)
}
\arguments{
\item{n}{The number of trials to simulate.}
}
\value{
Returns a data frame with columns 'rt' and 'correct'.
}
\description{
Perform a very quick simulation of one subject using dRiftDM.
}
