% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{r_from_sds}
\alias{r_from_sds}
\title{Correlation from Standard Deviations}
\usage{
r_from_sds(b, w)
}
\arguments{
\item{b}{numeric // The known between-subjects standard deviation.}

\item{w}{numeric // The known within-subjects standard deviation.}
}
\value{
numeric // The correlation of parameters.
}
\description{
Calculate the correlation of parameters given a known between- and within-subjects standard deviation.
}
