% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs.R
\name{rcs}
\alias{rcs}
\title{Rate correct score}
\usage{
rcs(rt, pc, n)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}

\item{n}{Integer indicating the number of trials that make up this block.}
}
\value{
A vector of scores of the same length as the passed vectors.
}
\description{
Calculate RCS for a given block. Each number should represent one
participants average across trials. RCS can be expressed formally as,
\deqn{RCS = \frac{NC}{\sum{RT}}}
}
