% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_variables.R
\name{remove_variables}
\alias{remove_variables}
\title{Remove All Variables}
\usage{
remove_variables()
}
\description{
Remove all objects from the global environment except constants. A constant is any object with an entirely uppercase name.
}
