% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_wrapper_functions.R
\name{return_bis_info}
\alias{return_bis_info}
\title{Return BIS Info}
\usage{
return_bis_info(.data, pars = list("drift", "boundary"), verbose = F)
}
\arguments{
\item{.data}{The data object to calculate BIS from.}

\item{pars}{The between-subjects parameters to calculate BIS across.}

\item{verbose}{Should stuff be printed to the console or not?}
}
\description{
Return a data frame of important BIS calculation.
}
