% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_wrapper_functions.R
\name{return_bis_list}
\alias{return_bis_list}
\title{Return BIS List.}
\usage{
return_bis_list(.data)
}
\arguments{
\item{.data}{data.frame to act on.}
}
\value{
Returns a list of parameters, with the absolute value of the bis score across the other parameter levels.
}
\description{
BIS wrapper for 2024 within-subjects simulation project.
}
