% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_wrapper_functions.R
\name{return_bis_matrix}
\alias{return_bis_matrix}
\title{Return BIS Matrix}
\usage{
return_bis_matrix(.data)
}
\arguments{
\item{.data}{data.frame to act on.}
}
\value{
Returns a matrix of BIS values across possible parameter combinations.
}
\description{
BIS wrapper for 2024 within-subjects simulation project.
}
