% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_wrapper_functions.R
\name{return_bis_vector}
\alias{return_bis_vector}
\title{Return BIS Vector.}
\usage{
return_bis_vector(.data)
}
\arguments{
\item{.data}{data.frame to act on.}
}
\value{
Returns a vector of length nrow(data).
}
\description{
BIS wrapper for 2024 within-subjects simulation project.
}
