% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{return_data_from_jatos}
\alias{return_data_from_jatos}
\title{Jatos to Data}
\usage{
return_data_from_jatos(
  path,
  subject_col,
  outcome_col,
  rt_col,
  block_col = "",
  deadline_col = "",
  difficulty_col = "",
  group_col = "",
  sat_col = "",
  sat_conditions = list(),
  trial_col = ""
)
}
\arguments{
\item{path}{Path to the folder where the JATOS file is stored. Used to find the first available file with 'jatos_results' in the title.}

\item{subject_col}{The current name of the subject column.}

\item{outcome_col}{The current name of the outcome / accuracy column.}

\item{rt_col}{The current name of the response time column.}

\item{block_col}{Optional. The current name of the block number or id column.}

\item{deadline_col}{Optional. The Current name of the deadline column.}

\item{difficulty_col}{Optional. The current name of the difficulty column.}

\item{group_col}{Optional. The current name of the between-subjects ID column.}

\item{sat_col}{Optional. The current name of the speed-accuracy within-subjects ID column.}

\item{sat_conditions}{Optional. List to rename SAT conditions from their internal names to more useful ones. Must be listed in format 'current'='new'.}

\item{trial_col}{Optional. The current name of the trial number column.}
}
\value{
A data frame with columns Subject, Outcome, and RT, which optimally includes Group, Block, Trial, SAT, Deadline, Difficulty.
}
\description{
Convert the messy JATOS csv file to a standardized data frame.
}
