% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{return_input}
\alias{return_input}
\title{Return an Input From User.}
\usage{
return_input(options, prompt = "Please choose from the following options:")
}
\arguments{
\item{options}{The type that the response must be given in.}

\item{prompt}{The prompt for the user, printed as a message.}
}
\value{
Returns one of the options provided to the options argument.
}
\description{
Give the user a set of options, each associated with a number, and get their input.
}
