% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{return_rounded_seconds}
\alias{return_rounded_seconds}
\title{Return Rounded Seconds from Milliseconds}
\usage{
return_rounded_seconds(milliseconds, dp = 3)
}
\arguments{
\item{milliseconds}{Data to convert to seconds. Expects numeric.}

\item{dp}{Optional. Number of decimal places to round to. Defaults to 3.}
}
\value{
A data frame of rounded seconds of the same type as the input data.
}
\description{
Divide milliseconds by 1000 and then round to specified decimal places.
}
