% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_wrapper_functions.R
\name{return_tts}
\alias{return_tts}
\title{Return T-Tests.}
\usage{
return_tts(.data)
}
\arguments{
\item{.data}{data.frame to act on.}
}
\value{
Returns a matrix of bools showing if the outcome of the t-test was significant or not.
}
\description{
T-Test wrapper for 2024 within-subjects simulation project.
}
