% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{rr}
\alias{rr}
\title{Reward rate (Trial)}
\usage{
rr(rt, nc)
}
\arguments{
\item{rt}{Vector of response times, one per trial.}

\item{nc}{Vector indicating the outcome of the trial.}
}
\value{
Vector of length(rt) indicating the reward rate.
}
\description{
Calculate reward rate (Gold & Shadlen, 2002) for each provided trial.
}
