% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr.R
\name{rr_block}
\alias{rr_block}
\title{Reward rate (Block)}
\usage{
rr_block(rt, pc, n)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}

\item{n}{Integer indicating the number of trials that make up this block.}
}
\value{
A vector of scores of the same length as the passed vectors.
}
\description{
Estimate reward rate (Gold & Shadlen, 2002) for the given averages.
}
