% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat_performance}
\alias{sat_performance}
\title{SAT Performance}
\format{
## `sat_performance`
A 60 x 3 tibble data.frame with columns boundary (dlb), rt (dbl), and pc (dbl).
}
\usage{
sat_performance
}
\description{
A simple data frame containing simulated data revealing
range of human performance across boundary conditions.
Boundary here is in DMCfun scale (sigma=4).
}
\keyword{datasets}
