% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{sd.cond_from_r}
\alias{sd.cond_from_r}
\title{SD(Cond) from Correlation}
\usage{
sd.cond_from_r(subject.sd, r)
}
\arguments{
\item{r}{numeric // The desired correlation.}

\item{sd}{numeric // The known standard deviation, either between- or within-subjects.}
}
\value{
numeric // The unknown standard deviation.
}
\description{
Calculate the condition-level standard deviation required to reach the desired correlation.
}
