% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{sd.subject_from_r}
\alias{sd.subject_from_r}
\title{SD(Subject) from Correlation}
\usage{
sd.subject_from_r(sd, r)
}
\arguments{
\item{sd}{numeric // The known standard deviation, either between- or within-subjects.}

\item{r}{numeric // The desired correlation.}
}
\value{
numeric // The unknown standard deviation.
}
\description{
Calculate the subject-level standard deviation required to reach the desired correlation.
}
