% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cell_colour.R
\name{set_cell_colour}
\alias{set_cell_colour}
\title{Set Cell Colour}
\usage{
set_cell_colour(table, x, y, colour, ignore_headings = T)
}
\arguments{
\item{table}{gridExtra::tableGrob // The table to set the colour of.}

\item{x}{numeric // The x position (row) to look at.}

\item{y}{numeric // The y position (column) to look at.}

\item{colour}{colour // The colour to set the background colour to.}

\item{ignore_headings}{bool // Should the x and y coordinates be adjusted to
ignore the heading row and row-number coolumn?}
}
\value{
gridExtra::tableGrob // Return the input table with the colour of the cell adjusted.
}
\description{
Set the background colour of the specified cell.
}
