% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridExtra.R
\name{set_cell_gpar_values}
\alias{set_cell_gpar_values}
\title{Set Cell Graphical Parameter Values}
\usage{
set_cell_gpar_values(table, x, y, ..., ignore_headings = T)
}
\arguments{
\item{table}{gridExtra::tableGrob // The table to set the colour of.}

\item{x}{numeric // The x position (row) to look at.}

\item{y}{numeric // The y position (column) to look at.}

\item{...}{List of arguments to pass to grid::gpar().}

\item{ignore_headings}{bool // Should the x and y coordinates be adjusted to
ignore the heading row and row-number coolumn?}
}
\value{
gridExtra::tableGrob // Return the input table with gpar adjustments applied.
}
\description{
Set the specific graphical parameters of the specified cell.
}
