% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_wrappers.R
\name{shelve_grobs}
\alias{shelve_grobs}
\title{Shelve Graphical Objects}
\usage{
shelve_grobs(..., heading = "", fontsize = 24)
}
\arguments{
\item{...}{list // List of grobs (or grob-like) objects to arrange.}

\item{heading}{string // Optional heading to append to shelved grobs.}

\item{fontsize}{numeric // Size of font used for heading.}
}
\value{
grid // gridExtra grid of shelve grobs with heading.
}
\description{
Arrange grobs horizontally and add a heading.
}
