% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\encoding{UTF-8}
\name{spinner}
\alias{spinner}
\title{Spinner}
\usage{
spinner(i, set = "circle", colour = "white")
}
\arguments{
\item{i}{Numeric variable indicating the current state of the loop.}

\item{set}{Symbol set to use for spinner. The following sets are currently available:
\itemize{
\item  circle:   ◐, ◓, ◑, ◒
\item  corner:   ▙, ▛, ▜, ▟
\item  quadrant: ▖, ▘, ▝, ▗
}}

\item{colour}{The colour to print the spinner in. Must be available in \code{\link{ansi_colours}}.}
}
\description{
cat a symbol to show the script is working. Always uses a \\r call to clear the line and replace the current percentage.
}
