% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{srt}
\alias{srt}
\title{Signed Residual Time (Trial)}
\usage{
srt(rt, x, deadline)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{x}{Numeric Vector of scores, should be coded correct = 1, incorrect = 0.}

\item{deadline}{Number or numeric vector of length(rt) indicating the trial deadline.}
}
\description{
Calculate signed residual time (Maris & van der Maas, 2012) for each provided trial.
}
