% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{srt_block}
\alias{srt_block}
\title{Signed Residual Time (Block)}
\usage{
srt_block(rt, pc, n, deadline)
}
\arguments{
\item{rt}{Float indicating the mean response time.}

\item{pc}{Float indicating the proportion of correct responses.}

\item{n}{Intiger indicating the number of trials that make up this block.}

\item{deadline}{Number indicating the mean trial deadline.}
}
\description{
Estimate signed residual time (Maris & van der Maas, 2012) for the given averages.
Useful when data is already aggregates across trials, but less accurate.
}
