% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssd.R
\name{ssd}
\alias{ssd}
\title{Sample Standard Deviation}
\usage{
ssd(sample, places = 3)
}
\arguments{
\item{sample}{vector-like. The sample from which to calculate the standard deviation.}

\item{places}{numeric, default = 3. Value to round to. Pass in 0 or FALSE to not round.}
}
\description{
Calculate the sample standard deviation.
}
