% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_grobs.R
\name{stack_grobs}
\alias{stack_grobs}
\title{Stack Graphical Objects}
\usage{
stack_grobs(..., heading = "", fontsize = 18)
}
\arguments{
\item{...}{list // List of grobs (or grob-like) objects to arrange.}

\item{heading}{string // Optional heading to append to stacked grobs.}

\item{fontsize}{numeric // Size of font used for heading.}
}
\value{
grid // gridExtra grid of stacked grobs with heading.
}
\description{
Arrange grobs vertically and add a heading.
}
