% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomR-package.R
\docType{package}
\name{tomR-package}
\alias{tomR}
\alias{tomR-package}
\title{tomR: Tom R}
\description{
A personal R package to reduce the amount of times I need to write the exact same lines of code.
}
\keyword{internal}
