% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{translate_threshold}
\alias{translate_threshold}
\title{Translate Threshold}
\usage{
translate_threshold(t, ms_to_s = T, dps = 3, quiet = F)
}
\arguments{
\item{t}{numeric. The original threshold to be converted.}

\item{ms_to_s}{bool, default = true. If true convert from ms to seconds.}

\item{dps}{numeric, default = 3. Value to round to. Pass in 0 or FALSE to not round.}

\item{quiet}{bool, default = false. If quiet do not print stuff to console.}
}
\description{
Translate threshold with sigma = 4 in ms to sigma = 1 in seconds.
}
