% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{bis}
\alias{bis}
\title{Balanced integration score}
\usage{
bis(rt, pc)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}
}
\description{
Calculate BIS () for a given block.
}
