% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{bis}
\alias{bis}
\title{Balanced integration score}
\usage{
bis(rt, pc)
}
\arguments{
\item{rt}{vector-like. The response time vector.}

\item{pc}{vector-like. The 'percent correct' vector.}
}
\description{
Calculate BIS.
}
