% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{cohens_d}
\alias{cohens_d}
\title{Cohen's D}
\usage{
cohens_d(x, y, paired = F)
}
\arguments{
\item{x}{Numeric vector of data from population x.}

\item{y}{Numeric vector of data from population y.}

\item{paired}{Bool indicating if the populations are paired or independent.
Paired populations have Cohen's recommended correction for Pearson's r automatically applied.}
}
\value{
The calculated d parameter, automatically correct for Pearson's r when required.
}
\description{
Calculate Cohen's d from the provided variables.
}
