% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{git_it}
\alias{git_it}
\title{Git It!}
\usage{
git_it(repo, quiet = F)
}
\arguments{
\item{repo}{The name of the Github repository to load from, as if using install_github. If already installed will use the package portion to load the library.}

\item{quiet}{Stop the function from printing messages to your console.}
}
\description{
Try to load a package you got from github, or, install it from github if you do not have it yet. If you do not have devtools the function will first install that. After doing all that, spin up the library in the current environment.
}
