% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{ies}
\alias{ies}
\title{Inverse efficiency score}
\usage{
ies(rt, pc)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}
}
\description{
Calculate IES.
}
