% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{ies}
\alias{ies}
\title{Inverse efficiency score}
\usage{
ies(rt, pc)
}
\arguments{
\item{rt}{vector-like. The response time vector.}

\item{pc}{vector-like. The 'percent correct' vector.}

\item{places}{numeric, default = 3. Value to round to. Pass in 0 or FALSE to not round.}
}
\description{
Calculate IES for a provided set of rts and pcs.
}
