% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ies.R
\name{ies}
\alias{ies}
\title{Inverse efficiency score}
\usage{
ies(rt, pc)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}
}
\value{
A vector of scores of the same length as the passed vectors.
}
\description{
Calculate IES for a given block. Each number should represent one
participants average across trials. IES can be expressed formally as,
\deqn{IES = \frac{\overline{RT}}{PC}}
}
