% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{lai}
\alias{lai}
\title{Log A Index (Trial)}
\usage{
lai(rt, p = 0.5, s = 0.2, A = log(999))
}
\arguments{
\item{rt}{Numeric vector of response times, one per trial.}

\item{p}{The probability of a correct response.}

\item{s}{The minimum response time required for accuracy to be greater than chance.}

\item{A}{Asymptotic accuracy, aka, the maximum value logit(p) can ever take.}
}
\description{
Calculate Log A Index (Dennis & Evans, 1996) for each provided trial.
}
