% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveload.R
\name{lifesaver}
\alias{lifesaver}
\title{Lifesaver}
\usage{
lifesaver(.object, file)
}
\arguments{
\item{.object}{The R object to try and save.}

\item{file}{The file path you want to save to.}
}
\value{
True if file was saved, false otherwise.
}
\description{
Wrapper to prevent accidentally saving over a file that already exists.
}
