% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisas.R
\name{lisas}
\alias{lisas}
\title{Linear integrated speed-accuracy score}
\usage{
lisas(rt, pc)
}
\arguments{
\item{rt}{Numeric vector of response times.}

\item{pc}{Numeric vector of proportion correct scores.}
}
\value{
A vector of scores of the same length as the passed vectors.
}
\description{
Calculate LISAS for a given block. Each number should represent
one participants average across trials.
\deqn{LISAS = \overline{RT} + \frac{S_{RT}}{S_{PE}}\cdot PE}
}
\note{
LISAS is vulnerable to failure when PE = 0 because we both require the
standard deviation fo the PE distribution, and we multiply by PE. Because LISAS
is a measure of RT adjusted for errors, when there are no errors there is nothing
to adjust for and we can instead use the following formula,
\deqn{LISAS = \overline{RT}}
}
