% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM.R
\name{non_dec_fixed}
\alias{non_dec_fixed}
\title{Fixed Non-Decision Time}
\usage{
non_dec_fixed(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{list // One row of prms_matrix or df_prms.}

\item{prms_solve}{list // The parameters relevant for deriving the PDFs.}

\item{t_vec}{list // Time space, from 0 to t_max with length nt + 1.}

\item{one_cond}{single character string // Indicates the current condition.}

\item{ddm_opts}{Additional arguments to pass to deeper dRiftDM calls.}
}
\value{
A vector of length t_vec containing non_dec from prms_model.
}
\description{
Function to calculate non-decision time when it is fixed across conditions.
}
