% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{psd}
\alias{psd}
\title{Percentage Standard Deviation}
\usage{
psd(value, dps = 3, quiet = F)
}
\arguments{
\item{value}{numeric. Percentage from which the SD should be calculated.}

\item{dps}{numeric, default = 3. Value to round to. Pass in 0 or FALSE to not round.}

\item{quiet}{bool, default = false. If quiet do not print stuff to console.}
}
\description{
Calculate the standard deviation of a percentage/mean value.
}
