% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_from_sds.R
\name{r_from_sds}
\alias{r_from_sds}
\title{Correlation from Standard Deviations}
\usage{
r_from_sds(subject.sd, cond.sd)
}
\arguments{
\item{subject.sd}{numeric // The known between-subjects standard deviation.}

\item{cond.sd}{numeric // The known within-subjects standard deviation.}
}
\value{
numeric // The correlation of parameters.
}
\description{
Calculate the correlation of parameters given a known between- and within-subjects standard deviation.
}
