% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{remove_variables}
\alias{remove_variables}
\title{Remove All Variables}
\usage{
remove_variables(exceptions = c(), verbose = F)
}
\arguments{
\item{exceptions}{A list of variables to be excepted from the rm command.}

\item{verbose}{List a bunch of stuff from the environment.}
}
\description{
Remove all objects from the global environment except constants. A constant is any object with an entirely uppercase name.
}
