% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{rr}
\alias{rr}
\title{Reward rate}
\usage{
rr(rt, pc)
}
\arguments{
\item{rt}{vector-like. The response time vector.}

\item{pc}{vector-like. The 'percent correct' vector.}
}
\description{
Calculate reward rate (Gold & Shadlen, 2002) for a provided set of rts and pcs.
}
