% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{rr_block}
\alias{rr_block}
\title{Reward rate (Block)}
\usage{
rr_block(rt, pc, n)
}
\arguments{
\item{rt}{Mean response time.}

\item{pc}{Proportion of correct responses.}

\item{n}{The number of trials that make up this block.}
}
\value{
Vector of length(rt) indicating the reward rate.
}
\description{
Estimate reward rate (Gold & Shadlen, 2002) for the given averages.
}
