% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_from_sds.R
\name{sd.cond_from_r}
\alias{sd.cond_from_r}
\title{SD(Cond) from Correlation}
\usage{
sd.cond_from_r(sd, r)
}
\arguments{
\item{sd}{numeric // The known standard deviation, either between- or within-subjects.}

\item{r}{numeric // The desired correlation.}
}
\value{
numeric // The unknown standard deviation.
}
\description{
Calculate the condition-level standard deviation required to reach the desired correlation.
}
