% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{sd_from_r}
\alias{sd_from_r}
\title{Standard Deviation from Correlation}
\usage{
sd_from_r(sd, r)
}
\arguments{
\item{sd}{numeric // The known standard deviation, either between- or within-subjects.}

\item{r}{numeric // The desired correlation.}
}
\value{
numeric // The unknown standard deviation.
}
\description{
Calculate the between- or within-subjects standard deviation required to
 reach the desired correlation. This works because r = b^2 / (b^2 + w^2), meaning
 we can solve for either b or w to determine the desired standard deviation.
}
