% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{srt}
\alias{srt}
\title{Signed Residual Time}
\usage{
srt(rt, pc)
}
\arguments{
\item{rt}{vector-like. The response time vector.}

\item{pc}{vector-like. The 'percent correct' vector.}
}
\description{
Calculate signed residual time (Maris & van der Maas, 2012) for a provided set of rts and pcs.
}
