% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_sat_measures.R
\name{srt}
\alias{srt}
\title{Signed Residual Time}
\usage{
srt(score, deadline, rt)
}
\arguments{
\item{score}{Numeric Vector // A vector of scores representing correct or incorrect.}

\item{deadline}{Numeric Vector // A vector of deadline times, i.e. the time at which a response must have been given by.}

\item{rt}{Numeric Vector // A vector of response times.}
}
\description{
Calculate signed residual time (Maris & van der Maas, 2012).
}
