% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{translate_drift}
\alias{translate_drift}
\title{Translate Drift Rate}
\usage{
translate_drift(d, dps = 3, ms_to_s = T, verbose = F)
}
\arguments{
\item{d}{numeric. The original drift rate to be converted.}

\item{dps}{numeric, default = 3 // Value to round to. Pass in 0 or FALSE to not round.}

\item{ms_to_s}{bool, default = true // If true convert from ms to seconds.}

\item{verbose}{bool, default = false // Should stuff be printed to the console or not?}
}
\description{
Translate drift rate with sigma = 4 in ms to sigma = 1 in seconds.
}
