% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_load.R
\name{try_load}
\alias{try_load}
\title{Try and Load a File.}
\usage{
try_load(path, error_message = "File not found.")
}
\arguments{
\item{path}{The path to the file you want to load.}

\item{error_message}{Message to display if an error occurs.}
}
\description{
Attempt to load a file, or stop and warn if the path is incorrect.
}
