% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{try_load}
\alias{try_load}
\title{Try and Load a File.}
\usage{
try_load(path, error_message)
}
\arguments{
\item{path}{The path to the file you want to load.}

\item{error_message}{The message to display if the file is not found.}
}
\description{
Attempt to load a file, or stop and warn if the path is incorrect.
}
